Process HW jobs
